requirements.txt
absl-py==1.3.0
appnope==0.1.3
arcade==2.6.16
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
branca==0.6.0
cachetools==5.2.1
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.7
entrypoints==0.4
executing==1.1.0
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.37.3
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
gridworld==0.1
grpcio==1.51.1
gym==0.26.1
gym-gridworlds==0.0.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.13.1
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.38.1
lxml==4.9.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.6.3
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
plotting==0.0.7
progressbar==2.5
prompt-toolkit==3.0.31
protobuf==4.25.1
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.2
pyglet==2.0.dev23
Pygments==2.13.0
pymunk==6.2.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytiled-parser==2.2.0
pytz==2022.2.1
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.8.1
seaborn==0.12.0
simpy==4.0.1
six==1.16.0
stack-data==0.5.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
# tensorflow==2.13.0
tensorflow-estimator==2.13.0
# tensorflow-macos==2.13.0
termcolor==2.2.0
threaded==4.1.0
threadpoolctl==3.3.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.8.1
